backend/requirements.txt
# Core Framework
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.6.0
Flask-Bcrypt==1.0.1

# Database
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9

# AI/ML Libraries
numpy==1.26.2
pandas==2.1.4
scikit-learn==1.3.2
tensorflow==2.15.0
torch==2.1.2
torchvision==0.16.2

# OCR and Image Processing
pytesseract==0.3.10
easyocr==1.7.1
Pillow==10.1.0
opencv-python==4.8.1.78

# NLP for Ingredient Analysis
spacy==3.7.2
transformers==4.36.2
nltk==3.8.1

# API and Validation
marshmallow==3.20.1
python-dotenv==1.0.0
requests==2.31.0

# Payment Processing
stripe==7.9.0
paypalrestsdk==1.13.1

# Utilities
python-dateutil==2.8.2
pytz==2023.3
gunicorn==21.2.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-flask==1.3.0

# Security
cryptography==41.0.7
